+++ test/libsolidity/syntaxTests/constructor/function_named_constructor.sol
@@ -1,0 +1,5 @@
+contract C {
+    function constructor() public;
+}
+// ----
+// Warning: (17-47): This function is named "constructor" but is not the constructor of the contract.